use newer types, add update function to ctx
