+++ src/components/DataNotPersistentNotification/stylesheet.scss
@@ -1,6 +1,6 @@
 @import '../../variables';
 
-.data-not-persistent-notification{
+.data-not-persistent-notification {
   font-size: 0.9rem;
   margin: 0;
   text-align: left;
