+++ src/contexts/schedule.ts
@@ -1,17 +1,22 @@
 import React from 'react';
+import { Draft, Immutable } from 'immer';
 
 import { Oscar } from '../data/beans';
 import { EMPTY_OSCAR } from '../data/beans/Oscar';
+import { defaultSchedule, Schedule } from '../data/types';
 import { ErrorWithFields } from '../log';
-import { TermData, defaultTermData } from '../types';
 
-export type ScheduleContextData = {
-  term: string;
-  oscar: Oscar;
-} & TermData;
+export type ScheduleContextData = Immutable<Schedule> & {
+  readonly term: string;
+  readonly oscar: Oscar;
+};
+
 export type ScheduleContextSetters = {
   setTerm: (next: string) => void;
-  patchSchedule: (patch: Partial<TermData>) => void;
+  patchSchedule: (patch: Partial<Schedule>) => void;
+  updateSchedule: (
+    applyDraft: (draft: Draft<Schedule>) => void | Immutable<Schedule>
+  ) => void;
 };
 export type ScheduleContextValue = [
   ScheduleContextData,
@@ -21,7 +26,7 @@ export const ScheduleContext = React.createContext<ScheduleContextValue>([
   {
     term: '',
     oscar: EMPTY_OSCAR,
-    ...defaultTermData,
+    ...defaultSchedule,
   },
   {
     setTerm: (next: string): void => {
@@ -32,12 +37,17 @@ export const ScheduleContext = React.createContext<ScheduleContextValue>([
         },
       });
     },
-    patchSchedule: (patch: Partial<TermData>): void => {
+    patchSchedule: (patch: Partial<Schedule>): void => {
       throw new ErrorWithFields({
         message: 'empty ScheduleContext.patchSchedule value being used',
         fields: {
           patch,
         },
+      });
+    },
+    updateSchedule: (): void => {
+      throw new ErrorWithFields({
+        message: 'empty ScheduleContext.updateSchedule value being used',
       });
     },
   },
