+++ src/data/legacy/types.ts
@@ -1,0 +1,15 @@
+export type TermData = {
+  desiredCourses: string[];
+  pinnedCrns: string[];
+  excludedCrns: string[];
+  colorMap: Record<string, string>;
+  sortingOptionIndex: number;
+};
+
+export const defaultTermData: TermData = {
+  desiredCourses: [],
+  pinnedCrns: [],
+  excludedCrns: [],
+  colorMap: {},
+  sortingOptionIndex: 0,
+};
